new timeline interface: dark blue: keyframe blue: selected frame white: frame orange: last frame of the movie
+ Auto toKeyframe system, automatically converts non-empty frames to keyframes

+++ trunk/Animator/freemovie/clearPreviews.php
@@ -1,11 +1,13 @@
 <?php
+	echo "emptying previews...";
 if ( $dh = opendir ( "preview/" ) ) {
 	while ( false !== ( $dat = readdir ( $dh ) ) ) {
-		if ( $dat != "." && $dat != ".." && $dat != "Thumbs.db" ) {
+		if ( $dat != "." && $dat != ".." && $dat != "Thumbs.db"  && $dat != ".svn") {
 			echo "'$dat' ,";
 			unlink("preview/$dat");
 		}
 	}
+	echo "done.";
 	closedir ( $dh );
 }
 ?>
